Submitted by:  Kirk McKusick
Allow chown() to return success if the gid isn't changed even if
the gid is not the caller's.  Such gids are normal for files created
in world-writable directories sucj as /tmp.  This "fixes" annoying
error messages for mv'ing files created in /tmp to another file
system.  mv still preserves the foreign gid of /tmp, but now does
it silently.