MFC:
"Add AlphaServer 2000 (demi-sable), 2100 (sable), and 2100A (lynx) support.

Only PCI and on-board ISA peripherials are supported at this time.
This support has been only lightly tested due to a lack of response to my
call for testers on the freebsd-alpha mailing list.  It works quite well
on the one AS2100 on which it has been tested, but it may not work on
an AS2100A and should therefore be regarded as experimental."

While this may be considered experimental, this should and is being
moved back to the stable branch so there is at least something to work
with before the tree stumps start being dynamited as part of the SMP
rototillage.