MFC mountd.c/1.67 and exports.5/1.20:
Give a more explanatory error message for the frequent pilot error of
attempting to export the non-root of a filesystem with the -alldirs
option.

Also merge the -quiet option that completely discards this error
message upon requesting so, so you can export some mount point with
-alldirs without getting complaints if there's nothing mounted at it
(e. g. like /cdrom).

Approved by:	re